src/shared/config/mongodb/requirements.txt
pymongo
python-dotenv
dnspython